Modification as discussed in Gitter
Default size of the app should be - 757, 341 (w, h)
When maximizing/fullscreen the window the lineedit size changes (it should stay the same).
Needs to set a minimum size (original way) for the lineedit.
when dragging the window it should change cursor.
When the Updater is poping-up it will be aligned to the center of the app window.