fix(ci): include runtime dlls individually
The previous approach wouldn't make the CI fail if some DLL was missing.
The current approach makes it more obvious.

Additionally, make sure to also include `vcruntime140_1.dll` on the x64 build.
This dll is not required to run the x86 build.

This was all tested on a clean Windows 7 64-bit virtual machine.