chore(build): Capitalize the project's name
Shouldn't make much of a difference, but since the QMake project was
stylized like that, I decided to do the same here.

The side effect is having the final executable called "Notes" instead
of "notes", which shouldn't matter on Windows because of
case-insensitive file names.

On macOS, because we're bundling the app into a .dmg, CMake will take
care of adjusting the executable name in Info.plist automatically.

The only exception to this is Linux, where it's conventional to have
the executable name in lowercase, so I added an exception to cover that.