chore(ci): Revert to CMake's Makefile generator on Linux
I'd love to have consistent build steps across all OSes, i.e.:

1. cmake . -B build -G Ninja
2. cmake --build build
3. cmake --install build

But in reality, this is rather annoying to accomplish on older (yet
still supported) Linux distros, which are likely not to have a version
of CMake recent enough to supports both --build and --install.

So it's easier to just go back to the good ol' make && make install,
which is also solid.

We can hopefully revisit this once Ubuntu 18.04 reaches end of life...