feat(build): Add GIT_REVISION option to CMake
It's disabled by default, but when enabled, it will try to append the
current commit's short hash to the app's version string, which makes it
easier to differentiate between builds while we're testing them.

This option is mostly intended for developers, which is why this commit
also enables it by default in our CI pipeline.