fix(listview): Don't move a note to the top if it's already there
Whenever a note was updated (i.e. by typing into it), we would move
that note to the very top of the list view.

The problem is that at every character typed, we trigger
ListViewLogic::moveNoteToTop(), which would then move that note to the
top, and would also update our 'Settings.ini', in order to change its
'currentSelectNotesId' property.

It might not look like it, but this whole process is very I/O heavy.
And while it was not noticeable on Linux/macOS, on Windows, however,
this would absolutely tank the performance (especially on slow storage).

To prevent a note from being moved to the top unnecessarily, we now
simply check if it's already there, therefore preventing Settings.ini
from also being updated unnecessarily.