feat(setup): Automatically uninstall different versions on Windows
If one or more different versions of Notes are currently installed, we
now tell users they must uninstall them before proceeding.

This behavior is enabled by default, but it can by completely bypassed
by invoking the installer via the command line with the
/SKIPVERSIONCHECK parameter.

That's mainly done because it's been reported in the past that some dlls
included in previous versions of the installer would cause issues.

And because these dlls don't get removed automatically upon installing a
new version, they could still cause issues...

We could also approach this problem by maintaining a list of problematic
files in [InstallDelete] [1], but in my opinion, it's much easier to
just uninstall any version that's different from the one the user is
trying to currently install.

[1] https://jrsoftware.org/ishelp/index.php?topic=installdeletesection