chore(ui): Remove system tray context menu on macOS
This is necessary because we want to show/hide the main window upon
clicking the system tray icon. However, on macOS especifically, we
can't do this without _also_ showing/hiding the system tray context
menu.

There's probably a better way to do this, but currently Qt APIs don't
seem to provide one.