fix: Ignore some events when using native window decorations
Currently, these events are only being used to handle window management
(moving, resizing, etc.) when we're in 'frameless mode'.

So, they can be ignored when the native window frame is used.

This fixes some mildly annoying bugs present when we're using native
window decorations, such as:

- Being able to maximize/unmaximize the window by double clicking next
to its title bar
- Being able to move or resize the window by grabbing next to its edges