fix: Forward AppKit's hide/unhide events to Qt
Closing the window on macOS merely hides the app to the dock. However,
because Qt is never informed of this, it will still think its window is
visible when we try to restore it, requiring the user to click on the
tray icon twice, for example.

We fix this by monitoring AppKit's hide/unhide notifications (events),
and then we simply call Qt's hide() / show() methods, accordingly.