fix: Forward Qt's hide/show events to AppKit
Clicking on the tray icon will call Qt's hide() / show() methods, but
AppKit is never informed of this.

So, we listen to Qt's showEvent / hideEvent, and then we simply call
AppKits' hide: / unhide: methods, accordingly.

Unfortunately, this does not seem to work every time, for some reason.

But I'd argue that's still better than not doing anything...