fix(ui): Update window margins on Linux on maximize/unmaximize
Now that we delegate window moving to the window manager via
startSystemMove(), some window managers on Linux allow us to snap the
window to the top edge of the screen, in order to maximize it.

Since we didn't know when this would happen, we were not removing the
margins added around the window (when in 'frameless mode'), causing
them to stay visible even when the window was maximized.