fix(ux): Only allow moving the window by grabbing its title bar
Previously, while in frameless mode, we could move the window by
grabbing some unusual places. Now, we copy what most window managers
do and only allow moving the window by grabbing the title bar.

The height of the title bar is defined by the height of the
verticalSpacer_upTreeView widget (currently 25 pixels), while the width
is just the width of the main window (corrected for the margins we add
around the window on Linux).