chore(build): Only enable extra warnings on MSVC 14.20 and newer
Apparently, on older MSVC versions (such as the one included in Visual
Studio 2017), the compiler flag to mark third-party sources as
"external" (an useful option to supress warnings on code we have no
control over), was still experimental and CMake didn't use it.

This change supresses compiler warnings altogether, while building on
Windows 7 with Visual Studio 2017, for example.