chore(build): Make CPack strip debug symbols
We were manually removing debug symbols from Linux binaries in CI, and
while that was working just fine, the introduction of the Kanban view
feature seems to have broken that, for some reason.

The side effect of that, is that running lintian on the deb package
would return an error.

Ideally, we should investigate the root cause of this, but for now this
seems like an acceptable solution.