fix(build): Add qt6-sql-sqlite as dependency for the RPM package
Unfortunately, on openSUSE Leap 15.4 the SQLite library we need is not
a hard dependency of the base Qt 6 package, like it is in some other
popular distros.

To work around this, we have to manually add qt6-sql-sqlite as a
dependency of the RPM package here.

Note: This is only done when we detect the current distro is openSUSE.

My manual test case:

1. Run openSUSE 15.4 (choosing the GNOME ISO is recommended, as it
   doesn't come with any Qt dependency pre-installed) in 'live mode'
2. Download the appropriate RPM package of Notes
3. Install the package with:
   sudo zypper install --allow-unsigned-rpm notes.rpm
4. Run 'notes' from the terminal

Before this commit, this error would show up and the app wouldn't work
correctly:

QSqlDatabase: SQLITE driver not loaded
QSqlDatabase: available drivers:
Error: connection with database fail
QSlQuery::exec database not open