Linux: Fix Wayland window associations
Wayland compositors use the desktopFileName window property to associate native Wayland windows with their .desktop launcher.

This was previously set to "notes" when qt launched as a native Wayland window (the default in Wayland sessions), which could cause various inconsistencies.

Notably, KDE Plasma would display the generic Wayland icon in task view for Notes instead of the icon referenced in the .desktop file.

Fix this by propagating the APP_ID property (which is used to name the .desktop file during packaging) to main.cpp so that we can call setDesktopFileName with it.

After this change KDE Plasma correctly displays the Notes icon in task view.