chore(ci): Build snap package
Everytime we push to the main branch (master), Snapcraft automatically
triggers a new build using _their_ build system, which is better than
GitHub Actions because it supports many CPU architectures.

However, we still have to make sure the snap package continues to
work, and that's why this commit adds a new CI job for the snap package
(amd64 only).