chore(ci): Build/publish Docker images used to create DEB packages
From now on, whenever we detect a modification to any of the 'ubuntu-*'
Dockerfiles in the 'Dockerfiles' directory, a new image will be built
first (with 'docker build'), and only then it will be used to actually
build the app and generate DEB packages.

If no changes to these Dockerfiles are detected, we'll simply
'docker pull' the already-existing image from the GitHub Container
Registry to build the app, which will give us a faster build time.

This new approach allows us to immediately test changes made to these
Dockerfiles, without having to first build them locally, test them,
push them to an OCI registry, and only then use them on GitHub Actions.

Please note that these images will only be pushed to the GitHub Container
Registry if _all_ three conditions below are met:

- We _are not_ on a pull request
- We _are_ on the master branch
- We _are_ on the 'nuttyartist/notes' repository