fix(ci): Update strategy that checks if a Dockerfile was modified
GitHub Actions gives us special variables that can be used to get the
last previous ref before the current one.

This works great for 'push' and 'pull_request' events, but we don't have
something similar for 'workflow_dispatch' events.

In the latter case, we just assume the last previous ref is the HEAD of
the 'master' branch.

Fixes #696