fixup!(ci): Run create-dmg until it succeeds
This properly stores the result of the increment operation in a
variable, where previously that evaluation could return 1 and mess up
with the exit status of this build step.

That would happen when create-dmg failed in the first try, but succeeded
in the next one.

I've also added "set -x" to enable Bash's command tracing, making it
easier to spot how many times create-dmg gets invoked.