Remove Qt 5 support
As time goes by, it becomes more and more laborious to maintain two
major Qt versions in parallel.

We had many workarounds and missing features for the Qt 5 version, which
had its last public release almost 4 years ago.

Additionally, the system requirements for Qt 6 are not unreasonable, so
the majority of people should be able to run it.

So, for those reasons, we decided to remove Qt 5 support completely.

Users that can't run the Qt 6 version for whatever reason, are free to
either disable updates and stay on Notes 2.3.x, or manually build Notes
off of the "qt5" branch.

Co-authored-by: guihkx <626206+guihkx@users.noreply.github.com>